+++ src/components/Modals/index.ts
@@ -1,7 +1,8 @@
 export { HelpModal } from './HelpModal';
 export { ImportModal } from './ImportModal';
-export { ShareModal } from './ShareModal';
 export { BinListModal } from './BinListModal';
 // Re-export from new location for backward compatibility
 // TODO: Remove in Phase 6 - update all imports to use features/print-export
 export { PrintModal } from '../../features/print-export';
+// TODO: Remove in Phase 6 - update all imports to use features/cloud-share
+export { ShareModal } from '../../features/cloud-share';
