+++ src/features/cloud-share/components/SharedLayoutImporter.tsx
@@ -1,17 +1,17 @@
 import { useEffect, useState, useCallback, useRef } from 'react';
-import { useLayoutStore } from '../../core/store/layout';
-import { useUIStore } from '../../core/store/ui';
-import { useHistoryStore } from '../../core/store/history';
-import { useToastStore } from '../../core/store/toast';
-import { useLibraryStore, computePreview } from '../../core/store/library';
+import { useLayoutStore } from '../../../core/store/layout';
+import { useUIStore } from '../../../core/store/ui';
+import { useHistoryStore } from '../../../core/store/history';
+import { useToastStore } from '../../../core/store/toast';
+import { useLibraryStore, computePreview } from '../../../core/store/library';
 import {
   getSharedLayoutFromURL,
   clearSharedLayoutFromURL,
   getCloudShareIdFromURL,
-} from '../../core/storage';
-import { fetchShare } from '../../core/api/share';
-import { isOk, getUserMessage } from '../../core/result';
-import type { Layout, SharePermission, LayoutPreview } from '../../core/types';
+} from '../../../core/storage';
+import { fetchShare } from '../../../core/api/share';
+import { isOk, getUserMessage } from '../../../core/result';
+import type { Layout, SharePermission, LayoutPreview } from '../../../core/types';
 
 // Check for shared layout once at module load time (URL-encoded shares)
 const initialShareResult = getSharedLayoutFromURL();
