+++ src/test/components/ShareModal.test.tsx
@@ -1,21 +1,18 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
-import { ShareModal } from '../../components/Modals/ShareModal';
+import { ShareModal } from '../../components/Modals';
 import { useLayoutStore, useLibraryStore, useUIStore, useLabsStore } from '../../core/store';
 import { resetAllStores } from '../testUtils';
 import * as storage from '../../core/storage';
 import * as analytics from '../../utils/analytics';
 
 // Mock CloudShareTab since it's a complex component
-vi.mock('../../components/Share', () => ({
+vi.mock('../../features/cloud-share/components/CloudShareTab', () => ({
   CloudShareTab: ({ onSwitchToUrlTab }: { onSwitchToUrlTab: () => void }) => (
     <div data-testid="cloud-share-tab">
       <button onClick={onSwitchToUrlTab}>Switch to URL</button>
     </div>
   ),
-  ShareButton: () => null,
-  SharedLayoutBanner: () => null,
-  SharedLayoutImporter: () => null,
 }));
 
 // Mock storage utilities
